refactor(oci): move OCI package from internal/oci to pkg/oci for external consumption
Move tag resolution, credential keychains, OCI client, and all
supporting types (errors, bundle serialization, cache, resolver) from
internal/oci/ to pkg/oci/ so that external consumers like the
pacto-operator can import github.com/trianalab/pacto/pkg/oci directly,
eliminating duplicated code. Update all internal import paths and
architecture documentation accordingly.
