+++ internal/app/resolve.go
@@ -10,9 +10,9 @@ import (
 	"path/filepath"
 	"testing/fstest"
 
-	"github.com/trianalab/pacto/internal/oci"
 	"github.com/trianalab/pacto/pkg/contract"
 	"github.com/trianalab/pacto/pkg/graph"
+	"github.com/trianalab/pacto/pkg/oci"
 	"github.com/trianalab/pacto/pkg/override"
 	"github.com/trianalab/pacto/pkg/validation"
 )
