+++ internal/cli/commands_test.go
@@ -12,9 +12,9 @@ import (
 
 	"github.com/trianalab/pacto/internal/app"
 	"github.com/trianalab/pacto/internal/cli"
-	"github.com/trianalab/pacto/internal/oci"
 	"github.com/trianalab/pacto/internal/testutil"
 	"github.com/trianalab/pacto/pkg/contract"
+	"github.com/trianalab/pacto/pkg/oci"
 	"github.com/trianalab/pacto/pkg/plugin"
 )
