+++ internal/cli/dashboard.go
@@ -13,8 +13,8 @@ import (
 	"github.com/spf13/cobra"
 	"github.com/spf13/viper"
 	"github.com/trianalab/pacto/internal/app"
-	"github.com/trianalab/pacto/internal/oci"
 	"github.com/trianalab/pacto/pkg/dashboard"
+	"github.com/trianalab/pacto/pkg/oci"
 )
 
 func newDashboardCommand(svc *app.Service, v *viper.Viper, version string) *cobra.Command {
