+++ pkg/dashboard/resolve_e2e_test.go
@@ -12,8 +12,8 @@ import (
 	"testing/fstest"
 	"time"
 
-	"github.com/trianalab/pacto/internal/oci"
 	"github.com/trianalab/pacto/pkg/contract"
+	"github.com/trianalab/pacto/pkg/oci"
 )
 
 // pullCountingStore tracks pull calls and returns a configurable bundle.
